# src/v2/docs/openapi.yaml
openapi: 3.0.0
info:
  title: TMJApp API V2
  version: 2.0.0
  description: API v2 organizada por visoes (admin, passenger, driver).
servers:
  - url: http://localhost:3000/api/v2
  - url: http://api.tmjapp.com.br/api/v2
...
paths:
...
  /driver/dashboard:
    get:
      tags: [Driver V2]
      summary: Dados consolidados do dashboard do motorista
      parameters:
        - $ref: '#/components/parameters/DriverHighPriorityFilter'
        - $ref: '#/components/parameters/DriverCategoryFilter'
        - $ref: '#/components/parameters/DriverLatFilter'
        - $ref: '#/components/parameters/DriverLngFilter'
        - $ref: '#/components/parameters/DriverRadiusKmFilter'
      responses:
        '200':
          description: Dashboard retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverDashboardResponse'
        '404':
          description: Motorista nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/availability:
    patch:
      tags: [Driver V2]
      summary: Alternar motorista online/offline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverAvailabilityRequest'
      responses:
        '200':
          description: Disponibilidade atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverAvailabilityResponse'
        '400':
          description: Availability invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Motorista nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/location:
    post:
      tags: [Driver V2]
      summary: Atualizar localizacao operacional do motorista online
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverOperationalLocationRequest'
      responses:
        '200':
          description: Localizacao atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverOperationalLocationResponse'
        '401':
          description: Token ausente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Motorista nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '422':
          description: Payload de localizacao invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '429':
          description: Janela minima entre atualizacoes nao respeitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/device-token:
    post:
      tags: [Driver V2]
      summary: Registrar token de push do dispositivo do motorista
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverDeviceTokenRequest'
      responses:
        '200':
          description: Token registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '401':
          description: Token ausente ou invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Motorista nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '422':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/pending:
    get:
      tags: [Driver V2]
      summary: Listar corridas pendentes com filtros operacionais
      parameters:
        - $ref: '#/components/parameters/DriverHighPriorityFilter'
        - $ref: '#/components/parameters/DriverCategoryFilter'
        - $ref: '#/components/parameters/DriverLatFilter'
        - $ref: '#/components/parameters/DriverLngFilter'
        - $ref: '#/components/parameters/DriverRadiusKmFilter'
        - $ref: '#/components/parameters/DriverLimitFilter'
        - $ref: '#/components/parameters/DriverOffsetFilter'
      responses:
        '200':
          description: Corridas pendentes listadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverPendingRidesResponse'
  /driver/rides/{rideId}/accept:
    post:
      tags: [Driver V2]
      summary: Aceitar corrida pendente (operacao atomica)
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Corrida aceita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverAcceptRideResponse'
        '400':
          description: Ride ID invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Corrida nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Corrida ja aceita ou indisponivel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Motorista offline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /driver/rides/{rideId}/realtime-token:
    post:
      tags: [Driver V2]
      summary: Gerar token realtime Firebase para tracking da corrida
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token realtime gerado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRealtimeTokenResponse'
        '400':
          description: Ride ID invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Corrida nao atribuida ao motorista autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Corrida nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/rides/{rideId}/location/snapshot:
    post:
      tags: [Driver V2]
      summary: Persistir snapshot oficial da localizacao do motorista
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverLocationSnapshotRequest'
      responses:
        '201':
          description: Snapshot salvo com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverLocationSnapshotResponse'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Corrida nao atribuida ao motorista autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Corrida nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          description: Snapshot enviado antes da janela minima
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/rides/{rideId}/eta:
    get:
      tags: [Driver V2]
      summary: Calcular ETA operacional com base no ultimo snapshot
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ETA retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideEtaResponse'
        '400':
          description: Ride ID invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Corrida nao atribuida ao motorista autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Corrida nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/rides/history/summary:
    get:
      tags: [Driver V2]
      summary: Resumo financeiro e operacional do histórico de corridas
      parameters:
        - in: query
          name: period
          required: false
          schema:
            type: string
            enum: [week, month, custom]
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Resumo retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideHistorySummaryResponse'
        '422':
          description: Faixa de datas inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/history:
    get:
      tags: [Driver V2]
      summary: Histórico paginado de corridas finalizadas do motorista
      parameters:
        - in: query
          name: period
          required: false
          schema:
            type: string
            enum: [week, month, custom]
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Histórico retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideHistoryListResponse'
        '422':
          description: Faixa de datas inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/history-detail:
    get:
      tags: [Driver V2]
      summary: Detalhe completo da corrida finalizada
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhe retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideHistoryDetail'
        '422':
          description: Ride ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/support/options:
    get:
      tags: [Driver V2]
      summary: Opções de suporte disponíveis para a corrida
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Opções retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideSupportOptionsResponse'
        '422':
          description: Ride ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/support/tickets:
    post:
      tags: [Driver V2]
      summary: Abrir ticket de suporte para a corrida
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRideSupportTicketRequest'
      responses:
        '201':
          description: Ticket criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideSupportTicketResponse'
        '422':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/support/payment-issue:
    post:
      tags: [Driver V2]
      summary: Relatar problema com valor recebido
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRidePaymentIssueRequest'
      responses:
        '201':
          description: Ticket criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideSupportSimpleTicketResponse'
        '422':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/support/forgotten-object:
    post:
      tags: [Driver V2]
      summary: Relatar objeto esquecido
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRideForgottenObjectRequest'
      responses:
        '201':
          description: Ticket criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRideSupportSimpleTicketResponse'
        '422':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
  /driver/rides/{rideId}/support/passenger-absent:
    post:
      tags: [Driver V2]
      summary: Solicitar cancelamento sem taxa por passageiro ausente
      parameters:
        - in: path
          name: rideId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverRidePassengerAbsentRequest'
      responses:
        '201':
          description: Solicitação aberta com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRidePassengerAbsentResponse'
        '422':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '404':
          description: Corrida não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
...
  /driver/wallet/summary:
    get:
      tags: [Driver V2]
      summary: Resumo da carteira do motorista
      parameters:
        - in: query
          name: period
          required: false
          schema:
            type: string
            enum: [today, week, month]
      responses:
        '200':
          description: Resumo retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverWalletSummaryResponse'
  /driver/wallet/activities:
    get:
      tags: [Driver V2]
      summary: Listar atividades/lançamentos da carteira
      parameters:
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [RIDE_CREDIT, BONUS, PIX_TRANSFER_DEBIT, ADJUSTMENT]
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Atividades retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverWalletActivitiesResponse'
  /driver/wallet/transfers/pix:
    post:
      tags: [Driver V2]
      summary: Solicitar transferência PIX (CPF)
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverPixTransferRequest'
      responses:
        '201':
          description: Transferência criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverPixTransferCreateResponse'
        '400':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Chave de idempotência em processamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '422':
          description: Saldo insuficiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/wallet/transfers/{transferId}:
    get:
      tags: [Driver V2]
      summary: Detalhes da transferência
      parameters:
        - in: path
          name: transferId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transferência retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverTransferDetailsResponse'
        '400':
          description: Transfer ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Transferência não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/wallet/transfers/{transferId}/receipt:
    get:
      tags: [Driver V2]
      summary: Retornar comprovante da transferência
      parameters:
        - in: path
          name: transferId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comprovante retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverTransferReceiptResponse'
        '400':
          description: Transfer ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Transferência não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/profile/photo:
    put:
      tags: [Driver V2]
      summary: Salvar foto de perfil do motorista autenticado
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Foto salva com sucesso e etapa de profile photo concluida
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/AuthenticatedUser'
        '400':
          description: Arquivo nao enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /driver/profile:
    get:
      tags: [Driver V2]
      summary: Buscar perfil do motorista autenticado
      responses:
        '200':
          description: Perfil retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverProfileResponse'
        '404':
          description: Motorista não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
    put:
      tags: [Driver V2]
      summary: Atualizar dados do perfil do motorista
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverProfileUpdateRequest'
      responses:
        '200':
          description: Perfil atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessageResponse'
        '409':
          description: E-mail ou telefone já utilizados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
        '422':
          description: Falha de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DriverHighPriorityFilter:
      in: query
      name: highPriority
      required: false
      schema:
        type: boolean
    DriverCategoryFilter:
      in: query
      name: category
      required: false
      schema:
        type: string
        enum: [delivery, passenger]
    DriverLatFilter:
      in: query
      name: lat
      required: false
      schema:
        type: number
        format: double
    DriverLngFilter:
      in: query
      name: lng
      required: false
      schema:
        type: number
        format: double
    DriverRadiusKmFilter:
      in: query
      name: radiusKm
      required: false
      schema:
        type: number
        format: double
...
  schemas:
...
    DriverAvailabilityRequest:
      type: object
      required: [availability]
      properties:
        availability:
          type: string
          enum: [ONLINE, OFFLINE]
...
    DriverDashboardResponse:
      type: object
      required: [success, availability, summary, pendingRides, currentRide]
      properties:
        success:
          type: boolean
        availability:
          type: string
          enum: [ONLINE, OFFLINE]
        summary:
          $ref: '#/components/schemas/DriverDashboardSummary'
        pendingRides:
          type: array
          items:
            $ref: '#/components/schemas/DriverPendingRideItem'
        currentRide:
          $ref: '#/components/schemas/DriverCurrentRide'
...
    ApiErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
        traceId:
          type: string
security:
  - bearerAuth: []